+++ WEB(FE)/mogoon/src/components/Header/HeaderItem.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+let HeaderItem =(props)=>{
+    return(
+        <li>{props.title}</li>
+    );
+};
+
+export default HeaderItem;
