+++ WEB(FE)/mogoon/src/index.css
@@ -1,3 +1,8 @@
+html,body{
+  height: 100%;
+  overflow: hidden;
+}
+
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
@@ -11,3 +16,9 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+ul,li{
+  list-style:none;
+  padding: 0px;
+}
+
